third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/get-element-by-id-in-shadow-mutation-expected.txt

CONSOLE ERROR: line 12: Uncaught TypeError: host.createShadowRoot is not a function
Tests to ensure that ShadowRoot.getElementById works even after mutation

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
